TT - added pretty time-based elf selection
