+++ src/lib/santa/SantaWait.svelte
@@ -1,0 +1,31 @@
+<script>
+    export let fullName, startDate, endDate
+</script>
+
+<div class="article-content mx-auto prose prose-headings:max-w-[56ch] prose-p:max-w-[56ch] text-prose text-justify text-zinc-900 dark:text-white dark:prose-invert prose-blockquote:not-italic prose-blockquote:font-handwriting prose-blockquote:border-amber-600 dark:prose-blockquote:border-amber-900">
+    <h2 class="text-lg mb-10">🎄✨ Ho, Ho, Ho! 🎅✨</h2>
+
+    <p class="text-lg mb-3">Congratulations, dear {fullName}, for joining our festive Secret Santa
+        game! 🎁
+        Your
+        cheerful spirit has brought even more warmth to our Christmas celebrations! 🌟 The twinkling lights,
+        the scent of gingerbread cookies, and the joyous laughter of friends and family all make this
+        holiday season truly magical! 🎉</p>
+
+    <p class="text-lg mb-3">Get ready to be part of the most wonderful exchange of surprises! Your
+        participation in our Secret Santa game will surely make someone's Christmas merry and bright! 🎅🎁
+        Remember, the spirit of giving is the most precious gift of all! 🌟</p>
+
+    <p class="text-lg mb-3">As the holiday countdown begins, we're thrilled to let you know that the drawing
+        of names for your Secret Santa exchange will take place between {startDate} and {endDate}. During
+        this time, the Christmas magic will be at its peak, ensuring that every name is matched
+        with care and excitement! 🎄✨</p>
+
+    <p class="text-lg mb-3">May your heart be filled with joy and kindness, and may your Secret Santa
+        experience be filled with delightful surprises and genuine smiles! 🌟 Wishing you a jolly and festive
+        holiday season, filled with love, laughter, and the spirit of giving! 🎅❤️</p>
+
+    <p class="text-lg mb-3">Happy Holidays and Merry Christmas! 🎄🎅🎁</p>
+
+    <p class="text-lg mb-3">Warmest wishes, St@k 🌟</p>
+</div>
