+++ src/lib/utils/index.js
@@ -1,20 +1,24 @@
 export const fetchMarkdownPosts = async () => {
-	const allPostFiles = import.meta.glob('/src/routes/blog/*.md')
-	const iterablePostFiles = Object.entries(allPostFiles)
+    const allPostFiles = import.meta.glob('/src/routes/blog/*.md')
+    const iterablePostFiles = Object.entries(allPostFiles)
 
-	const allPosts = await Promise.all(
-		iterablePostFiles.map(async ([path, resolver]) => {
-			const { metadata } = await resolver()
-			const postPath = path.slice(11, -3)
+    return await Promise.all(
+        iterablePostFiles.map(async ([path, resolver]) => {
+            const {metadata} = await resolver()
+            const postPath = path.slice(11, -3)
 
-			return {
-				meta: metadata,
-				path: postPath,
-			}
-		})
-	)
-
-	return allPosts
+            return {
+                meta: metadata,
+                path: postPath,
+            }
+        })
+    )
 }
 
-export const getRandomInt = (min, max) => (Math.floor(Math.pow(10, 14) * Math.random() * Math.random()) % (max - min + 1)) + min;
+export const isDateBetween = (startDate, endDate) => {
+    const currentDate = new Date()
+    return startDate <= currentDate && currentDate <= endDate
+}
+
+export const getRandomInt = (min, max) =>
+    (Math.floor(Math.pow(10, 14) * Math.random() * Math.random()) % (max - min + 1)) + min
